+++ CHANGELOG.md
@@ -8,12 +8,17 @@ and this project adheres to
 
 ## [Unreleased]
 
-## [0.2.0] - 2026-06-23
+## [0.3.0] - 2026-06-24
 
 ### Added
 
 - `pick` directory path argument
-- `start` alias
+- `start` alias for `init` subcommand
+
+## [0.2.0] - 2026-06-23
+
+### Added
+
 - `state` subcommand
 - `current` subcommand
 
@@ -38,6 +43,7 @@ and this project adheres to
 - `remove`/`rm` subcommand
 - `send` subcommand
 
-[unreleased]: https://github.com/rasmusmerzin/mpvd/compare/v0.2.0...main
+[unreleased]: https://github.com/rasmusmerzin/mpvd/compare/v0.3.0...main
+[0.2.0]: https://github.com/rasmusmerzin/mpvd/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/rasmusmerzin/mpvd/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/rasmusmerzin/mpvd/tree/v0.1.0